@@ -4,21 +4,17 @@
   "description": "a basic user app",
   "main": "./sdist/app.js",
   "scripts": {
+    "build-css": "node-sass-chokidar src/public/stylesheets -o sdist/public/stylesheets",
+    "watch-css": "npm run build-css && node-sass-chokidar src/public/stylesheets -o sdist/public/stylesheets --watch --recursive",
+    "watch-webpack": "webpack --watch",
     "build": "node ./sdist/config/build/build.js",
     "start": "node ./sdist/app.js",
     "test": "nyc mocha ./test/*",
-    "nuke": "psql postgres -a -f ./records/nuke.sql",
     "yarg": "node ./yarg-start.js"
   },
   "author": "thelionthatiam",
   "license": "ISC",
   "dependencies": {
-    "@types/bcrypt": "^1.0.0",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/filesystem": "0.0.28",
-    "@types/handlebars": "^4.0.36",
-    "@types/method-override": "0.0.31",
-    "@types/uuid": "^3.4.3",
     "an-array-of-english-words": "^1.3.1",
     "bcrypt": "^1.0.3",
     "body-parser": "^1.18.2",
@@ -28,6 +24,7 @@
     "file-system": "^2.2.2",
     "is-uuid": "^1.0.2",
     "method-override": "^2.3.10",
+    "node-sass-chokidar": "0.0.3",
     "nodemailer": "^4.1.0",
     "nodemailer-express-handlebars": "^2.0.0",
     "nodemailer-html-to-text": "^2.1.0",
@@ -35,34 +32,49 @@
     "pg": "^7.3.0",
     "promise": "^8.0.1",
     "prompt": "^1.0.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "request": "^2.83.0",
+    "runtime": "^0.14.1",
     "strict-mode": "^1.1.2",
     "uuid": "^3.1.0",
     "yargs": "^10.0.3"
   },
   "devDependencies": {
+    "@types/bcrypt": "^1.0.0",
     "@types/body-parser": "^1.16.7",
+    "@types/cookie-parser": "^1.4.1",
     "@types/express": "^4.0.39",
     "@types/express-handlebars": "0.0.29",
     "@types/express-session": "^1.15.5",
+    "@types/filesystem": "0.0.28",
+    "@types/handlebars": "^4.0.36",
     "@types/jquery": "^3.2.16",
     "@types/lodash": "^4.14.80",
+    "@types/method-override": "0.0.31",
     "@types/mocha": "^2.2.44",
     "@types/node": "^8.0.47",
     "@types/nodemailer": "^3.1.5",
     "@types/pg": "^7.1.5",
+    "@types/react": "^16.0.38",
+    "@types/react-dom": "^16.0.4",
+    "@types/uuid": "^3.4.3",
+    "awesome-typescript-loader": "^3.4.1",
     "chai": "^4.1.2",
     "chai-http": "^3.0.0",
     "istanbul": "^0.4.5",
     "mocha": "^4.0.1",
     "nyc": "^11.2.1",
+    "rollup": "^0.56.2",
+    "rollup-plugin-tsx": "^1.0.2",
+    "rollup-plugin-typescript": "^0.8.1",
     "superagent": "^3.8.0",
     "supertest": "^3.0.0",
     "supertest-session": "^3.1.0",
     "ts-loader": "^3.1.1",
     "tslint": "^5.8.0",
     "typescript": "^2.6.1",
-    "webpack": "^3.8.1"
+    "webpack": "^3.11.0"
   },
   "directories": {
     "test": "test"
