@@ -1,189 +1,228 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const isUUID = require("is-uuid");
-class UserException {
-    constructor(name, message) {
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = Object.setPrototypeOf ||
+        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import * as isUUID from 'is-uuid';
+var UserException = /** @class */ (function () {
+    function UserException(name, message) {
         this.message = message;
         this.name = name.toUpperCase();
     }
-}
-exports.UserException = UserException;
-class ValidationError extends Error {
-    constructor(name, message) {
-        super(message);
-        Error.captureStackTrace(this); // will creae a stack trace.
-        this.message = message;
-        this.name = name;
+    return UserException;
+}());
+var ValidationError = /** @class */ (function (_super) {
+    __extends(ValidationError, _super);
+    function ValidationError(name, message) {
+        var _this = _super.call(this, message) || this;
+        Error.captureStackTrace(_this); // will creae a stack trace.
+        _this.message = message;
+        _this.name = name;
+        return _this;
     }
-    static isOkay() {
+    ValidationError.isOkay = function () {
         return false;
-    }
-    toJSON() {
+    };
+    ValidationError.prototype.toJSON = function () {
         return {
             name: this.name,
             message: this.message
         };
-    }
-}
-exports.ValidationError = ValidationError;
-class ValueScalar {
-    constructor(value) {
+    };
+    return ValidationError;
+}(Error));
+var ValueScalar = /** @class */ (function () {
+    function ValueScalar(value) {
         this._value = value;
     }
-    valueOf() {
+    ValueScalar.prototype.valueOf = function () {
         return this._value;
-    }
-    toString() {
+    };
+    ValueScalar.prototype.toString = function () {
         return this._value;
-    }
-    toJSON() {
+    };
+    ValueScalar.prototype.toJSON = function () {
         return this._value;
+    };
+    return ValueScalar;
+}());
+var Email = /** @class */ (function (_super) {
+    __extends(Email, _super);
+    function Email() {
+        return _super !== null && _super.apply(this, arguments) || this;
     }
-}
-exports.ValueScalar = ValueScalar;
-class Email extends ValueScalar {
-    static fromJSON(value) {
+    Email.fromJSON = function (value) {
         return Email.create(value);
-    }
-    static create(value) {
-        let res = Email.validate(value);
+    };
+    Email.create = function (value) {
+        var res = Email.validate(value);
         if (!res.isOkay)
             throw res;
         return new Email(value);
-    }
-    static validate(email) {
-        let re = /^[A-Za-z0-9\._\$%\-]+@[A-Za-z0-9\-]+.[A-Za-z0-9]{2,6}$/;
+    };
+    Email.validate = function (email) {
+        var re = /^[A-Za-z0-9\._\$%\-]+@[A-Za-z0-9\-]+.[A-Za-z0-9]{2,6}$/;
         if (re.test(email)) {
             return { isOkay: true };
         }
         else {
             return new ValidationError('invalid type', 'This value -- ' + email + ' -- is not an email.');
         }
+    };
+    return Email;
+}(ValueScalar));
+var Permission = /** @class */ (function (_super) {
+    __extends(Permission, _super);
+    function Permission() {
+        return _super !== null && _super.apply(this, arguments) || this;
     }
-}
-exports.Email = Email;
-class Permission extends ValueScalar {
-    static fromJSON(value) {
+    Permission.fromJSON = function (value) {
         return Permission.create(value);
-    }
-    static create(value) {
-        let res = Permission.validate(value);
+    };
+    Permission.create = function (value) {
+        var res = Permission.validate(value);
         if (!res.isOkay)
             throw res;
         return new Permission(value);
-    }
-    static validate(permission) {
-        let re = /^(guest)$|^(user)$|^(admin)$/;
+    };
+    Permission.validate = function (permission) {
+        var re = /^(guest)$|^(user)$|^(admin)$/;
         if (re.test(permission)) {
             return { isOkay: true };
         }
         else {
             throw new ValidationError('invalid type', 'This value -- ' + permission + ' -- is not a permission.');
         }
+    };
+    return Permission;
+}(ValueScalar));
+var CharOnly = /** @class */ (function (_super) {
+    __extends(CharOnly, _super);
+    function CharOnly() {
+        return _super !== null && _super.apply(this, arguments) || this;
     }
-}
-exports.Permission = Permission;
-class CharOnly extends ValueScalar {
-    static fromJSON(value) {
+    CharOnly.fromJSON = function (value) {
         return CharOnly.create(value);
-    }
-    static create(value) {
-        let res = CharOnly.validate(value);
+    };
+    CharOnly.create = function (value) {
+        var res = CharOnly.validate(value);
         if (!res.isOkay)
             throw res;
         return new CharOnly(value);
-    }
-    static validate(charOnly) {
-        let re = /^[a-zA-Z ]+$/;
+    };
+    CharOnly.validate = function (charOnly) {
+        var re = /^[a-zA-Z ]+$/;
         if (re.test(charOnly)) {
             return { isOkay: true };
         }
         else {
             throw new ValidationError('invalid type', 'This value -- ' + charOnly + ' -- is not only chars.');
         }
+    };
+    return CharOnly;
+}(ValueScalar));
+var NumOnly = /** @class */ (function (_super) {
+    __extends(NumOnly, _super);
+    function NumOnly() {
+        return _super !== null && _super.apply(this, arguments) || this;
     }
-}
-exports.CharOnly = CharOnly;
-class NumOnly extends ValueScalar {
-    static fromJSON(value) {
+    NumOnly.fromJSON = function (value) {
         return NumOnly.create(value);
-    }
-    static create(value) {
-        let res = NumOnly.validate(value);
+    };
+    NumOnly.create = function (value) {
+        var res = NumOnly.validate(value);
         if (!res.isOkay)
             throw res;
         return new NumOnly(value);
-    }
-    static validate(numOnly) {
-        let re = /^[0-9.]+$/;
+    };
+    NumOnly.validate = function (numOnly) {
+        var re = /^[0-9.]+$/;
         if (re.test(numOnly)) {
             return { isOkay: true };
         }
         else {
             throw new ValidationError('invalid type', 'This value -- ' + numOnly + ' -- is not only numbers.');
         }
+    };
+    return NumOnly;
+}(ValueScalar));
+var UUID = /** @class */ (function (_super) {
+    __extends(UUID, _super);
+    function UUID() {
+        return _super !== null && _super.apply(this, arguments) || this;
     }
-}
-exports.NumOnly = NumOnly;
-class UUID extends ValueScalar {
-    static fromJSON(value) {
+    UUID.fromJSON = function (value) {
         return UUID.create(value);
-    }
-    static create(value) {
-        let res = UUID.validate(value);
+    };
+    UUID.create = function (value) {
+        var res = UUID.validate(value);
         if (!res.isOkay)
             throw res;
         return new UUID(value);
-    }
-    static validate(uuid) {
+    };
+    UUID.validate = function (uuid) {
         if (isUUID.v4(uuid)) {
             return { isOkay: true };
         }
         else {
             return new ValidationError('invalid type', 'This value -- ' + uuid + ' -- is not a uuid.');
         }
+    };
+    return UUID;
+}(ValueScalar));
+var Bool = /** @class */ (function (_super) {
+    __extends(Bool, _super);
+    function Bool() {
+        return _super !== null && _super.apply(this, arguments) || this;
     }
-}
-exports.UUID = UUID;
-class Bool extends ValueScalar {
-    static fromJSON(value) {
+    Bool.fromJSON = function (value) {
         return Bool.create(value);
-    }
-    static create(value) {
-        let res = Bool.validate(value);
+    };
+    Bool.create = function (value) {
+        var res = Bool.validate(value);
         if (!res.isOkay)
             throw res;
         return new Bool(value);
-    }
-    static validate(boolean) {
+    };
+    Bool.validate = function (boolean) {
         if (typeof boolean === 'boolean') {
             return { isOkay: true };
         }
         else {
             return new ValidationError('invalid type', 'This value -- ' + boolean + ' -- is not a boolean.');
         }
+    };
+    return Bool;
+}(ValueScalar));
+var String = /** @class */ (function (_super) {
+    __extends(String, _super);
+    function String() {
+        return _super !== null && _super.apply(this, arguments) || this;
     }
-}
-exports.Bool = Bool;
-class String extends ValueScalar {
-    static fromJSON(value) {
+    String.fromJSON = function (value) {
         return String.create(value);
-    }
-    static create(value) {
-        let res = String.validate(value);
+    };
+    String.create = function (value) {
+        var res = String.validate(value);
         if (!res.isOkay)
             throw res;
         return new String(value);
-    }
-    static validate(string) {
-        let re = /.+/;
+    };
+    String.validate = function (string) {
+        var re = /.+/;
         if (re.test(string)) {
             return { isOkay: true };
         }
         else {
             return new ValidationError('invalid type', 'This value -- ' + string + ' -- is not a string.');
         }
-    }
-}
-exports.String = String;
+    };
+    return String;
+}(ValueScalar));
+export { UserException, ValidationError, ValueScalar, Email, Permission, CharOnly, NumOnly, UUID, Bool, String, };
 //# sourceMappingURL=validation.js.map
