@@ -1,42 +1,40 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const express = require("express");
-const bcrypt = require("bcrypt");
-const async_database_1 = require("../middleware/async-database");
-const router = express.Router();
-let viewPrefix = 'account/';
+import * as express from 'express';
+import * as bcrypt from 'bcrypt';
+import { db } from '../middleware/async-database';
+var router = express.Router();
+var viewPrefix = 'account/';
 router.route('/:email')
-    .get((req, res) => {
+    .get(function (req, res) {
     res.render(viewPrefix + 'my-account', {
         email: req.session.user.email,
     });
 })
-    .delete((req, res) => {
-    async_database_1.db.query('DELETE FROM users WHERE user_uuid = $1', [req.session.user.uuid])
-        .then((result) => {
+    .delete(function (req, res) {
+    db.query('DELETE FROM users WHERE user_uuid = $1', [req.session.user.uuid])
+        .then(function (result) {
         res.render('login', {
             message: "account was deleted, please make a new one to enter"
         });
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render(viewPrefix + 'my-account', { dbError: err.stack });
     });
 });
 router.route('/:email/contact')
-    .get((req, res) => {
+    .get(function (req, res) {
     res.render(viewPrefix + 'my-contact', {
         email: req.session.user.email,
         phone: req.session.user.phone
     });
 })
-    .put((req, res) => {
-    let email = req.body.email;
-    let phone = req.body.phone;
-    let query = 'UPDATE users SET (email, phone) = ($1, $2) WHERE user_uuid = $3';
-    let input = [email, phone, req.session.user.uuid];
-    async_database_1.db.query(query, input)
-        .then((result) => {
+    .put(function (req, res) {
+    var email = req.body.email;
+    var phone = req.body.phone;
+    var query = 'UPDATE users SET (email, phone) = ($1, $2) WHERE user_uuid = $3';
+    var input = [email, phone, req.session.user.uuid];
+    db.query(query, input)
+        .then(function (result) {
         req.session.user.email = email;
         req.session.user.phone = phone;
         res.render(viewPrefix + 'my-account', {
@@ -44,28 +42,28 @@ router.route('/:email/contact')
             email: req.session.user.email
         });
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render(viewPrefix + 'my-account', { dbError: err.stack });
     });
 });
 router.route('/:email/password')
-    .get((req, res) => {
+    .get(function (req, res) {
     res.render(viewPrefix + 'new-password', {
         email: req.session.user.email
     });
 })
-    .post((req, res) => {
-    let inputs = {
+    .post(function (req, res) {
+    var inputs = {
         password: req.body.password,
         oldPassword: req.body.oldPassword
     };
-    async_database_1.db.query("SELECT * FROM users WHERE user_uuid = $1", [req.session.user.uuid])
-        .then((result) => {
+    db.query("SELECT * FROM users WHERE user_uuid = $1", [req.session.user.uuid])
+        .then(function (result) {
         console.log(result);
         return bcrypt.compare(req.body.oldPassword, result.rows[0].password);
     })
-        .then((result) => {
+        .then(function (result) {
         if (result === false) {
             throw new Error('Password incorrect');
         }
@@ -73,19 +71,19 @@ router.route('/:email/password')
             return bcrypt.hash(inputs.password, 10);
         }
     })
-        .then((hash) => {
+        .then(function (hash) {
         inputs.password = hash;
-        let query = 'UPDATE users SET password = $1 WHERE user_uuid = $2';
-        let input = [inputs.password, req.session.user.uuid];
-        return async_database_1.db.query(query, input);
+        var query = 'UPDATE users SET password = $1 WHERE user_uuid = $2';
+        var input = [inputs.password, req.session.user.uuid];
+        return db.query(query, input);
     })
-        .then((result) => {
+        .then(function (result) {
         res.render(viewPrefix + 'new-password', {
             success: true,
             email: req.session.user.email
         });
     })
-        .catch((error) => {
+        .catch(function (error) {
         console.log(error);
         res.render(viewPrefix + 'new-password', { dbError: error });
     });
