@@ -1,10 +1,8 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const express = require("express");
-const helper = require("../functions/helpers");
-const bcrypt = require("bcrypt");
-const database_1 = require("../middleware/database");
-const router = express.Router();
+import * as express from 'express';
+import * as helper from '../functions/helpers';
+import * as bcrypt from 'bcrypt';
+import { db } from '../middleware/database';
+var router = express.Router();
 //to sign up page
 router.get('/new-account', function (req, res, next) {
     res.render('new-account', { success: false });
@@ -15,8 +13,8 @@ router.post('/delete', function (req, res, next) {
     });
 });
 router.route('/accounts')
-    .post((req, res) => {
-    let inputs = {
+    .post(function (req, res) {
+    var inputs = {
         email: req.body.email,
         phone: req.body.phone,
         password: req.body.password,
@@ -25,38 +23,38 @@ router.route('/accounts')
         nonce: ''
     };
     bcrypt.hash(inputs.password, 10)
-        .then((hash) => {
+        .then(function (hash) {
         inputs.password = hash;
-        let query = 'INSERT INTO users(email, phone, password, name) VALUES($1, $2, $3, $4) RETURNING *';
-        let input = [inputs.email, inputs.phone, inputs.password, inputs.name];
-        return database_1.db.query(query, input);
+        var query = 'INSERT INTO users(email, phone, password, name) VALUES($1, $2, $3, $4) RETURNING *';
+        var input = [inputs.email, inputs.phone, inputs.password, inputs.name];
+        return db.query(query, input);
     })
-        .then((result) => {
+        .then(function (result) {
         inputs.uuid = result.rows[0].user_uuid;
         return helper.randomString;
     })
-        .then((string) => {
+        .then(function (string) {
         return bcrypt.hash(string, 10);
     })
-        .then((hash) => {
+        .then(function (hash) {
         inputs.nonce = hash;
-        return database_1.db.query('INSERT INTO nonce(user_uuid, nonce) VALUES ($1, $2) RETURNING *', [inputs.uuid, inputs.nonce]);
+        return db.query('INSERT INTO nonce(user_uuid, nonce) VALUES ($1, $2) RETURNING *', [inputs.uuid, inputs.nonce]);
     })
-        .then((result) => {
-        let query = 'INSERT INTO session (user_uuid, sessionID) VALUES ($1, $2)';
-        let input = [inputs.uuid, req.sessionID];
-        return database_1.db.query(query, input);
+        .then(function (result) {
+        var query = 'INSERT INTO session (user_uuid, sessionID) VALUES ($1, $2)';
+        var input = [inputs.uuid, req.sessionID];
+        return db.query(query, input);
     })
-        .then((result) => {
-        return database_1.db.query('INSERT INTO cart (user_uuid, card_number) VALUES ($1, $2)', [inputs.uuid, '42424242424242']);
+        .then(function (result) {
+        return db.query('INSERT INTO cart (user_uuid, card_number) VALUES ($1, $2)', [inputs.uuid, '42424242424242']);
     })
-        .then((result) => {
-        return database_1.db.query('INSERT INTO user_settings(user_uuid) VALUES ($1)', [inputs.uuid]);
+        .then(function (result) {
+        return db.query('INSERT INTO user_settings(user_uuid) VALUES ($1)', [inputs.uuid]);
     })
-        .then((result) => {
-        return database_1.db.query('UPDATE users SET permission = $1 WHERE user_uuid = $2', ['user', inputs.uuid]);
+        .then(function (result) {
+        return db.query('UPDATE users SET permission = $1 WHERE user_uuid = $2', ['user', inputs.uuid]);
     })
-        .then((result) => {
+        .then(function (result) {
         console.log(result.rows);
         res.render('new-account', {
             success: true,
@@ -64,27 +62,27 @@ router.route('/accounts')
             phone: inputs.phone,
         });
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err);
-        database_1.db.query('DELETE FROM users WHERE user_uuid = $1', [inputs.uuid])
-            .then((result) => {
-            return database_1.db.query('DELETE FROM nonce WHERE user_uuid = $1', [inputs.uuid]);
+        db.query('DELETE FROM users WHERE user_uuid = $1', [inputs.uuid])
+            .then(function (result) {
+            return db.query('DELETE FROM nonce WHERE user_uuid = $1', [inputs.uuid]);
         })
-            .then((result) => {
-            return database_1.db.query('DELETE FROM cart WHERE user_uuid = $1', [inputs.uuid]);
+            .then(function (result) {
+            return db.query('DELETE FROM cart WHERE user_uuid = $1', [inputs.uuid]);
         })
-            .then((result) => {
-            return database_1.db.query('DELETE FROM session WHERE user_uuid = $1', [inputs.uuid]);
+            .then(function (result) {
+            return db.query('DELETE FROM session WHERE user_uuid = $1', [inputs.uuid]);
         })
-            .then((result) => {
-            return database_1.db.query('DELETE FROM user_settings WHERE user_uuid = $1', [inputs.uuid]);
+            .then(function (result) {
+            return db.query('DELETE FROM user_settings WHERE user_uuid = $1', [inputs.uuid]);
         })
-            .then((result) => {
+            .then(function (result) {
             res.render('new-account', {
                 dbError: err.message
             });
         })
-            .catch((err) => {
+            .catch(function (err) {
             console.log(err);
             res.render('error');
         });
