@@ -1,49 +1,47 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const express = require("express");
-const database_1 = require("../../middleware/database");
-const router = express.Router();
+import * as express from 'express';
+import { db } from '../../middleware/database';
+var router = express.Router();
 //////////////
 ////////////// SHOW/CREATE USERS
 //////////////
 router.route('/accounts')
-    .get((req, res) => {
-    let accountContent;
-    let alarmContent;
-    let alarmArray = [];
-    let paymientContent;
-    let paymentArray = [];
-    let permission = false;
+    .get(function (req, res) {
+    var accountContent;
+    var alarmContent;
+    var alarmArray = [];
+    var paymientContent;
+    var paymentArray = [];
+    var permission = false;
     if (req.session.user.permission === 'admin') {
         permission = true;
     }
     else {
         permission = false;
     }
-    database_1.db.query("SELECT * FROM users", [])
-        .then((result) => {
+    db.query("SELECT * FROM users", [])
+        .then(function (result) {
         accountContent = result.rows;
-        for (let i = 0; i < accountContent.length; i++) {
-            alarmArray.push(database_1.db.query('SELECT * FROM alarms WHERE user_uuid = $1', [accountContent[i].user_uuid]));
+        for (var i = 0; i < accountContent.length; i++) {
+            alarmArray.push(db.query('SELECT * FROM alarms WHERE user_uuid = $1', [accountContent[i].user_uuid]));
         }
         return Promise.all(alarmArray);
     })
-        .then((result) => {
-        for (let i = 0; i < result.length; i++) {
+        .then(function (result) {
+        for (var i = 0; i < result.length; i++) {
             accountContent[i].alarmContent = result[i].rows;
-            for (let j = 0; j < accountContent[i].alarmContent.length; j++) {
+            for (var j = 0; j < accountContent[i].alarmContent.length; j++) {
                 accountContent[i].alarmContent[j].permission = permission;
             }
         }
-        for (let i = 0; i < accountContent.length; i++) {
-            paymentArray.push(database_1.db.query('SELECT * FROM payment_credit WHERE user_uuid = $1', [accountContent[i].user_uuid]));
+        for (var i = 0; i < accountContent.length; i++) {
+            paymentArray.push(db.query('SELECT * FROM payment_credit WHERE user_uuid = $1', [accountContent[i].user_uuid]));
         }
         return Promise.all(paymentArray);
     })
-        .then((result) => {
-        for (let i = 0; i < result.length; i++) {
+        .then(function (result) {
+        for (var i = 0; i < result.length; i++) {
             accountContent[i].paymentContent = result[i].rows;
-            for (let j = 0; j < accountContent[i].paymentContent.length; j++) {
+            for (var j = 0; j < accountContent[i].paymentContent.length; j++) {
                 accountContent[i].paymentContent[j].permission = permission;
                 console.log(accountContent[i].paymentContent[j].permission);
             }
@@ -52,7 +50,7 @@ router.route('/accounts')
             accountContent: accountContent,
         });
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err);
         res.render('error', {
             errName: err.message,
@@ -61,15 +59,15 @@ router.route('/accounts')
     });
 });
 router.route('/accounts/:user_uuid')
-    .delete((req, res) => {
-    let user_uuid = req.body.user_uuid;
-    let query = 'DELETE FROM users WHERE user_uuid = $1';
-    let input = [user_uuid];
-    database_1.db.query(query, input)
-        .then((result) => {
+    .delete(function (req, res) {
+    var user_uuid = req.body.user_uuid;
+    var query = 'DELETE FROM users WHERE user_uuid = $1';
+    var input = [user_uuid];
+    db.query(query, input)
+        .then(function (result) {
         res;
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
@@ -78,12 +76,12 @@ router.route('/accounts/:user_uuid')
 //////////// EDIT/DELETE USER CONTACT
 ////////////
 router.route('/accounts/:user_uuid/contact')
-    .get((req, res) => {
-    let user_uuid = req.query.user_uuid;
-    let permission = false;
-    database_1.db.query("SELECT * FROM users WHERE user_uuid = $1", [user_uuid])
-        .then((result) => {
-        let user = result.rows[0];
+    .get(function (req, res) {
+    var user_uuid = req.query.user_uuid;
+    var permission = false;
+    db.query("SELECT * FROM users WHERE user_uuid = $1", [user_uuid])
+        .then(function (result) {
+        var user = result.rows[0];
         if (req.session.user.permission === 'admin') {
             permission = true;
         }
@@ -97,31 +95,31 @@ router.route('/accounts/:user_uuid/contact')
             user_uuid: user_uuid
         });
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
 })
-    .put((req, res) => {
-    let email = req.body.a, phone = req.body.phone, user_uuid = req.body.user_uuid;
-    let query = 'UPDATE users SET (email, phone) = ($1, $2) WHERE user_uuid = $3';
-    let input = [email, phone, user_uuid];
-    database_1.db.query(query, input)
-        .then((result) => {
+    .put(function (req, res) {
+    var email = req.body.a, phone = req.body.phone, user_uuid = req.body.user_uuid;
+    var query = 'UPDATE users SET (email, phone) = ($1, $2) WHERE user_uuid = $3';
+    var input = [email, phone, user_uuid];
+    db.query(query, input)
+        .then(function (result) {
         res.redirect('/admin/accounts');
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
 })
-    .delete((req, res) => {
-    let user_uuid = req.body.user_uuid;
-    database_1.db.query('DELETE FROM users WHERE user_uuid = $1', [user_uuid])
-        .then((result) => {
+    .delete(function (req, res) {
+    var user_uuid = req.body.user_uuid;
+    db.query('DELETE FROM users WHERE user_uuid = $1', [user_uuid])
+        .then(function (result) {
         res.redirect('/admin/accounts');
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
@@ -130,15 +128,15 @@ router.route('/accounts/:user_uuid/contact')
 //////////// EDIT/DELETE USER ALARM
 ////////////
 router.route('/accounts/:user_uuid/alarms/:alarm_uuid')
-    .get((req, res) => {
+    .get(function (req, res) {
     console.log('ALARM ACCOUNT ROUTE');
-    let user_uuid = req.query.user_uuid;
-    let alarm_uuid = req.query.alarm_uuid;
-    let permission = false;
-    let alarm;
+    var user_uuid = req.query.user_uuid;
+    var alarm_uuid = req.query.alarm_uuid;
+    var permission = false;
+    var alarm;
     console.log(user_uuid, alarm_uuid, permission);
-    database_1.db.query("SELECT * FROM alarms WHERE user_uuid = $1 AND alarm_uuid = $2", [user_uuid, alarm_uuid])
-        .then((result) => {
+    db.query("SELECT * FROM alarms WHERE user_uuid = $1 AND alarm_uuid = $2", [user_uuid, alarm_uuid])
+        .then(function (result) {
         alarm = result.rows[0];
         console.log(alarm);
         if (req.session.user.permission === 'admin') {
@@ -156,36 +154,36 @@ router.route('/accounts/:user_uuid/alarms/:alarm_uuid')
             user_uuid: user_uuid
         });
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
 })
-    .put((req, res) => {
-    let user_uuid = req.body.user_uuid;
-    let alarm_uuid = req.body.alarm_uuid;
-    let title = req.body.title;
-    let awake = req.body.awake;
-    let active = req.body.active;
-    let query = 'UPDATE alarms SET (title, awake, active) = ($1, $2, $3) WHERE user_uuid = $4 AND alarm_uuid = $5';
-    let input = [title, awake, active, user_uuid, alarm_uuid];
-    database_1.db.query(query, input)
-        .then((result) => {
+    .put(function (req, res) {
+    var user_uuid = req.body.user_uuid;
+    var alarm_uuid = req.body.alarm_uuid;
+    var title = req.body.title;
+    var awake = req.body.awake;
+    var active = req.body.active;
+    var query = 'UPDATE alarms SET (title, awake, active) = ($1, $2, $3) WHERE user_uuid = $4 AND alarm_uuid = $5';
+    var input = [title, awake, active, user_uuid, alarm_uuid];
+    db.query(query, input)
+        .then(function (result) {
         res.redirect('/admin/accounts');
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
 })
-    .delete((req, res) => {
-    let user_uuid = req.body.user_uuid;
-    let alarm_uuid = req.body.alarm_uuid;
-    database_1.db.query('DELETE FROM alarms WHERE user_uuid = $1 AND alarm_uuid =$2', [user_uuid, alarm_uuid])
-        .then((result) => {
+    .delete(function (req, res) {
+    var user_uuid = req.body.user_uuid;
+    var alarm_uuid = req.body.alarm_uuid;
+    db.query('DELETE FROM alarms WHERE user_uuid = $1 AND alarm_uuid =$2', [user_uuid, alarm_uuid])
+        .then(function (result) {
         res.redirect('/admin/accounts');
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
@@ -194,14 +192,14 @@ router.route('/accounts/:user_uuid/alarms/:alarm_uuid')
 //////////// EDIT/DELETE USER PAYMENTS
 ////////////
 router.route('/accounts/:user_uuid/payment/:card_number')
-    .get((req, res) => {
-    let user_uuid = req.query.user_uuid;
-    let card_number = req.query.card_number;
-    let permission = false;
-    let payment;
+    .get(function (req, res) {
+    var user_uuid = req.query.user_uuid;
+    var card_number = req.query.card_number;
+    var permission = false;
+    var payment;
     console.log(user_uuid, card_number, permission);
-    database_1.db.query("SELECT * FROM payment_credit WHERE user_uuid = $1 AND card_number = $2", [user_uuid, card_number])
-        .then((result) => {
+    db.query("SELECT * FROM payment_credit WHERE user_uuid = $1 AND card_number = $2", [user_uuid, card_number])
+        .then(function (result) {
         payment = result.rows[0];
         console.log(payment);
         if (req.session.user.permission === 'admin') {
@@ -224,14 +222,14 @@ router.route('/accounts/:user_uuid/payment/:card_number')
             permission: permission
         });
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
 })
-    .put((req, res) => {
-    let oldCard = req.body.oldCard;
-    let inputs = {
+    .put(function (req, res) {
+    var oldCard = req.body.oldCard;
+    var inputs = {
         name: req.body.name,
         cardNumber: req.body.cardNumber,
         expDay: req.body.expDay,
@@ -242,28 +240,28 @@ router.route('/accounts/:user_uuid/payment/:card_number')
         state: req.body.state,
         zip: req.body.zip,
     };
-    let user_uuid = req.body.user_uuid;
-    let query = 'UPDATE payment_credit SET (card_number, name, exp_month, exp_date, cvv, address_1, city, state, zip) = ($1, $2, $3, $4, $5, $6, $7, $8, $9) WHERE user_uuid = $10 AND card_number = $11';
-    let input = [inputs.cardNumber, inputs.name, inputs.expMonth, inputs.expDay, inputs.cvv, inputs.address, inputs.city, inputs.state, inputs.zip, user_uuid, oldCard];
-    database_1.db.query(query, input)
-        .then((result) => {
+    var user_uuid = req.body.user_uuid;
+    var query = 'UPDATE payment_credit SET (card_number, name, exp_month, exp_date, cvv, address_1, city, state, zip) = ($1, $2, $3, $4, $5, $6, $7, $8, $9) WHERE user_uuid = $10 AND card_number = $11';
+    var input = [inputs.cardNumber, inputs.name, inputs.expMonth, inputs.expDay, inputs.cvv, inputs.address, inputs.city, inputs.state, inputs.zip, user_uuid, oldCard];
+    db.query(query, input)
+        .then(function (result) {
         res.redirect('/admin/accounts');
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
 })
-    .delete((req, res) => {
-    let user_uuid = req.body.user_uuid;
-    let card_number = req.body.card_number;
-    let query = 'DELETE FROM payment_credit WHERE user_uuid = $1 AND card_number =$2';
-    let input = [req.session.user.uuid, card_number];
-    database_1.db.query(query, input)
-        .then((result) => {
+    .delete(function (req, res) {
+    var user_uuid = req.body.user_uuid;
+    var card_number = req.body.card_number;
+    var query = 'DELETE FROM payment_credit WHERE user_uuid = $1 AND card_number =$2';
+    var input = [req.session.user.uuid, card_number];
+    db.query(query, input)
+        .then(function (result) {
         res.redirect('/admin/accounts');
     })
-        .catch((err) => {
+        .catch(function (err) {
         console.log(err.stack);
         res.render('error', { dbError: err.stack });
     });
