@@ -1,20 +1,18 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const express = require("express");
-const r = require("../resources/value-objects");
-const database_1 = require("../middleware/database");
-const EventEmitter = require('events').EventEmitter;
-const router = express.Router();
-const mockUser = {
+import * as express from 'express';
+import * as r from '../resources/value-objects';
+import { db } from '../middleware/database';
+var EventEmitter = require('events').EventEmitter;
+var router = express.Router();
+var mockUser = {
     uuid: '700fb850-ac21-4418-b843-8e4a0999bdf1',
     name: 'timothy',
     permission: 'user',
     cart_uuid: '70cf640a-cf9d-439c-98b5-0cfd8473ba02',
     email: 'g@g.gg'
 };
-const user = r.UserSession.fromJSON(mockUser);
-class Alarm {
-    constructor(user_uuid, title, active, alarm_uuid, state, time) {
+var user = r.UserSession.fromJSON(mockUser);
+var Alarm = /** @class */ (function () {
+    function Alarm(user_uuid, title, active, alarm_uuid, state, time) {
         this.user_uuid = user_uuid;
         this.title = title;
         this.active = active;
@@ -22,125 +20,131 @@ class Alarm {
         this.state = state;
         this.time = time;
     }
-    addSnooze() {
+    Alarm.prototype.addSnooze = function () {
+        var _this = this;
         console.log('------YOU SNOOZED! Now you have a snooze, but dont snooze to much!------');
-        let query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
-        let input = ['snoozing', this.user_uuid, this.alarm_uuid];
-        database_1.db.query(query, input)
-            .then((result) => {
-            return database_1.db.query('INSERT INTO snoozes(user_uuid, alarm_uuid) VALUES ($1, $2)', [this.user_uuid, this.alarm_uuid]);
+        var query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
+        var input = ['snoozing', this.user_uuid, this.alarm_uuid];
+        db.query(query, input)
+            .then(function (result) {
+            return db.query('INSERT INTO snoozes(user_uuid, alarm_uuid) VALUES ($1, $2)', [_this.user_uuid, _this.alarm_uuid]);
         })
-            .catch((error) => {
+            .catch(function (error) {
             console.log(error);
         });
-    }
-    addDismiss() {
+    };
+    Alarm.prototype.addDismiss = function () {
+        var _this = this;
         console.log('------YOU SLEPT IN! Now you have a dismiss under your belt.------');
-        let query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
-        let input = ['dismissed', this.user_uuid, this.alarm_uuid];
-        database_1.db.query(query, input)
-            .then((result) => {
-            return database_1.db.query('INSERT INTO dismisses(user_uuid, alarm_uuid) VALUES ($1, $2)', [this.user_uuid, this.alarm_uuid]);
+        var query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
+        var input = ['dismissed', this.user_uuid, this.alarm_uuid];
+        db.query(query, input)
+            .then(function (result) {
+            return db.query('INSERT INTO dismisses(user_uuid, alarm_uuid) VALUES ($1, $2)', [_this.user_uuid, _this.alarm_uuid]);
         })
-            .catch((error) => {
+            .catch(function (error) {
             console.log(error);
         });
-    }
-    addWake() {
+    };
+    Alarm.prototype.addWake = function () {
+        var _this = this;
         console.log('------NICE JOB, YOU WOKE UP! CARPE DIEM!------');
-        let query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
-        let input = ['woke', this.user_uuid, this.alarm_uuid];
-        database_1.db.query(query, input)
-            .then((result) => {
-            return database_1.db.query('INSERT INTO wakes(user_uuid, alarm_uuid) VALUES ($1, $2)', [this.user_uuid, this.alarm_uuid]);
+        var query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
+        var input = ['woke', this.user_uuid, this.alarm_uuid];
+        db.query(query, input)
+            .then(function (result) {
+            return db.query('INSERT INTO wakes(user_uuid, alarm_uuid) VALUES ($1, $2)', [_this.user_uuid, _this.alarm_uuid]);
         })
-            .catch((error) => {
+            .catch(function (error) {
             console.log(error);
         });
-    }
-    triggerAlarm() {
+    };
+    Alarm.prototype.triggerAlarm = function () {
         console.log('------ALARM TRIGGERED------');
-        let query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
-        let input = ['ringing', this.user_uuid, this.alarm_uuid];
-        database_1.db.query(query, input)
-            .then((result) => {
+        var query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
+        var input = ['ringing', this.user_uuid, this.alarm_uuid];
+        db.query(query, input)
+            .then(function (result) {
             console.log(result);
         })
-            .catch((error) => {
+            .catch(function (error) {
             console.log(error);
         });
-    }
-    alarmReset() {
-        let query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
-        let input = ['pending', this.user_uuid, this.alarm_uuid];
-        database_1.db.query(query, input)
-            .then((result) => {
+    };
+    Alarm.prototype.alarmReset = function () {
+        var query = 'UPDATE alarms SET state = $1 WHERE user_uuid = $2 AND alarm_uuid = $3';
+        var input = ['pending', this.user_uuid, this.alarm_uuid];
+        db.query(query, input)
+            .then(function (result) {
             console.log(result);
         })
-            .catch((error) => {
+            .catch(function (error) {
             console.log(error);
         });
-    }
-    snoozing() {
-        let currentTime = Date.now();
-        let endTime = currentTime + 10000; // ten second snooze
+    };
+    Alarm.prototype.snoozing = function () {
+        var _this = this;
+        var currentTime = Date.now();
+        var endTime = currentTime + 10000; // ten second snooze
         this.addSnooze();
-        let thing = setInterval(() => {
-            let timeLeft = endTime - Date.now();
+        var thing = setInterval(function () {
+            var timeLeft = endTime - Date.now();
             console.log('timeleft', timeLeft);
-            database_1.db.query('SELECT state FROM alarms WHERE user_uuid = $1 AND alarm_uuid = $2', [this.user_uuid, this.alarm_uuid])
-                .then((result) => {
-                let state = result.rows[0].state;
+            db.query('SELECT state FROM alarms WHERE user_uuid = $1 AND alarm_uuid = $2', [_this.user_uuid, _this.alarm_uuid])
+                .then(function (result) {
+                var state = result.rows[0].state;
                 console.log(state);
                 if (state === 'snoozing') {
                     if (timeLeft < 0) {
                         clearInterval(thing);
-                        this.triggerAlarm();
-                        this.ringing();
+                        _this.triggerAlarm();
+                        _this.ringing();
                     }
                 }
                 else if (state === 'dismiss') {
                     clearInterval(thing);
-                    this.addDismiss();
+                    _this.addDismiss();
                 }
                 else if (state === 'woke') {
                     clearInterval(thing);
-                    this.addWake();
+                    _this.addWake();
                 }
             });
         }, 1000);
-    }
-    ringing() {
-        let currentTime = Date.now();
-        let endTime = currentTime + 6000;
-        let thing = setInterval(() => {
-            let timeLeft = endTime - Date.now();
+    };
+    Alarm.prototype.ringing = function () {
+        var _this = this;
+        var currentTime = Date.now();
+        var endTime = currentTime + 6000;
+        var thing = setInterval(function () {
+            var timeLeft = endTime - Date.now();
             console.log('timeleft', timeLeft);
-            database_1.db.query('SELECT state FROM alarms WHERE user_uuid = $1 AND alarm_uuid = $2', [this.user_uuid, this.alarm_uuid])
-                .then((result) => {
-                let state = result.rows[0].state;
+            db.query('SELECT state FROM alarms WHERE user_uuid = $1 AND alarm_uuid = $2', [_this.user_uuid, _this.alarm_uuid])
+                .then(function (result) {
+                var state = result.rows[0].state;
                 console.log(state);
                 if (state === 'ringing') {
                     if (timeLeft < 0) {
                         clearInterval(thing);
-                        this.addDismiss();
+                        _this.addDismiss();
                     }
                 }
                 else if (state === 'snoozing') {
                     clearInterval(thing);
-                    this.snoozing();
+                    _this.snoozing();
                 }
                 else if (state === 'dismiss') {
                     clearInterval(thing);
-                    this.addDismiss();
+                    _this.addDismiss();
                 }
                 else if (state === 'woke') {
                     clearInterval(thing);
-                    this.addWake();
+                    _this.addWake();
                 }
             });
         }, 1000);
-    }
-}
+    };
+    return Alarm;
+}());
 module.exports = router;
 //# sourceMappingURL=alarm-class.js.map
