@@ -1,4 +1,4 @@
-const path = require('path');
+var path = require('path');
 module.exports = {
     entry: './src/index.ts',
     devtool: 'inline-source-map',
