@@ -1,33 +1,29 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-let test = [];
+var test = [];
 function invoiceItems(result) {
-    let invoiceItems = [];
-    let total = 0;
-    for (let i = 0; i < result.length; i++) {
-        let subtotal = result[i].price * result[i].quantity;
-        let quantity = result[i].quantity;
-        let name = result[i].name;
-        let obj = {
+    var invoiceItems = [];
+    var total = 0;
+    for (var i = 0; i < result.length; i++) {
+        var subtotal = result[i].price * result[i].quantity;
+        var quantity = result[i].quantity;
+        var name_1 = result[i].name;
+        var obj = {
             subtotal: subtotal,
             quantity: quantity,
-            name: name
+            name: name_1
         };
         invoiceItems.push(obj);
     }
     return invoiceItems;
 }
-exports.invoiceItems = invoiceItems;
 function total(invoiceItems) {
-    let total = 0;
-    for (let i = 0; i < invoiceItems.length; i++) {
+    var total = 0;
+    for (var i = 0; i < invoiceItems.length; i++) {
         total = total + invoiceItems[i].subtotal;
     }
     return total;
 }
-exports.total = total;
 function addDiscount(recieptContent) {
-    for (let i = 0; i < recieptContent.length; i++) {
+    for (var i = 0; i < recieptContent.length; i++) {
         if (recieptContent[i].discount === 0) {
             recieptContent[i].isDiscount = false;
         }
@@ -38,20 +34,38 @@ function addDiscount(recieptContent) {
     }
     return recieptContent;
 }
-exports.addDiscount = addDiscount;
 function addEmail(recieptContent, email) {
-    for (let i = 0; i < recieptContent.length; i++) {
+    for (var i = 0; i < recieptContent.length; i++) {
         recieptContent[i].email = email;
     }
     return recieptContent;
 }
-exports.addEmail = addEmail;
 function totalItems(recieptContent) {
-    let totalQuantity = 0;
-    for (let i = 0; i < recieptContent.length; i++) {
+    var totalQuantity = 0;
+    for (var i = 0; i < recieptContent.length; i++) {
         totalQuantity = totalQuantity + parseInt(recieptContent[i].quantity, 10);
     }
     return totalQuantity;
 }
-exports.totalItems = totalItems;
+// DUMMY DATA
+// let flan = [ {
+//     product_id: 'NRA0-S-GUNS-4233',
+//     name: 'National Rifle Association',
+//     price: '2.00',
+//     size: 's',
+//     description: 'Advocates for gun-owners rights in America.',
+//     quantity: '1',
+//     discount: 0,
+//     email: 'b@b.bb'
+//   },
+//    {
+//       product_id: 'PPH0-S-FEM0-5783',
+//       name: 'Planned Parenthood',
+//       price: '2.00',
+//       size: 's',
+//       description: 'Advocates for womens rights in America.',
+//       quantity: '25',
+//       discount: 0,
+//       email: 'b@b.bb' } ]
+export { invoiceItems, total, addDiscount, addEmail, totalItems };
 //# sourceMappingURL=invoice.js.map
