@@ -1,5 +1,3 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
 function isCyclic(obj) {
     var seenObjects = [];
     function detect(obj) {
@@ -21,7 +19,7 @@ function isCyclic(obj) {
 }
 function isScalar(obj) { return (/string|number|boolean/).test(typeof obj); }
 function mergeArray(arr1, arr2) {
-    let concatArray = arr1.concat(arr2);
+    var concatArray = arr1.concat(arr2);
     return concatArray;
 }
 function typeSort(item1, item2) {
@@ -57,7 +55,7 @@ function typeSort(item1, item2) {
 }
 function deepMerge(one, two) {
     if ((!isCyclic(one)) && (!isCyclic(two))) {
-        for (let prop in two) {
+        for (var prop in two) {
             one[prop] = typeSort(one[prop], two[prop]);
         }
     }
@@ -66,7 +64,6 @@ function deepMerge(one, two) {
     }
     return one;
 }
-exports.deepMerge = deepMerge;
 function MergeArray(array) {
     if (array.length === 0) {
         return 'error, no object in array';
@@ -75,10 +72,11 @@ function MergeArray(array) {
         return array[0];
     }
     else if (array.length >= 2) {
-        for (let i = 1; i < array.length; i++) {
-            let thisMerge = deepMerge(array[i], array[i - 1]);
+        for (var i = 1; i < array.length; i++) {
+            var thisMerge = deepMerge(array[i], array[i - 1]);
             array.pop();
         }
     }
 }
+export { deepMerge };
 //# sourceMappingURL=merge.js.map
