@@ -71,7 +71,7 @@ unsquishButton.click(function () {
 });
 // format date a little better
 $('.thedate').each(function () {
-    let dateChanger = /[a-zA-Z]{2,4} [a-zA-Z]{2,4} [\d]{1,2} [\d]{2,4}/g;
+    var dateChanger = /[a-zA-Z]{2,4} [a-zA-Z]{2,4} [\d]{1,2} [\d]{2,4}/g;
     var date = $(this).text();
     var newDateFormat = date.match(dateChanger);
     $(this).text(date.replace(date, ("created on: " + newDateFormat)));
