@@ -1,7 +1,8 @@
 // let delete = $('.delete-alarm');
-$('.delete-alarm').click(() => {
-    let id = $(this).attr('id');
-    $(this).toggleClass('red');
+var _this = this;
+$('.delete-alarm').click(function () {
+    var id = $(_this).attr('id');
+    $(_this).toggleClass('red');
     console.log(id);
 });
 //# sourceMappingURL=alarm-div-behavior.js.map
