@@ -5,15 +5,15 @@ function addLeadingZeros(number) {
     return number;
 }
 function theTime() {
-    let clock = new Date();
-    let h = addLeadingZeros(clock.getHours());
-    let m = addLeadingZeros(clock.getMinutes());
-    let s = addLeadingZeros(clock.getSeconds());
+    var clock = new Date();
+    var h = addLeadingZeros(clock.getHours());
+    var m = addLeadingZeros(clock.getMinutes());
+    var s = addLeadingZeros(clock.getSeconds());
     return h + ':' + m;
 }
 window.setInterval(function () {
-    let time = theTime();
-    let awakeTime = $('.awakeTime').text();
+    var time = theTime();
+    var awakeTime = $('.awakeTime').text();
     console.log(typeof awakeTime);
     $('#clock').text(time);
     if (time === awakeTime) {
